refactor: import other components async in component doc
